packages/infra/lib/NestedStacks/C2pa/Lambda/code/requirements.txt
boto3==1.37.5
fastapi==0.115.11
cryptography==44.0.2
c2pa-python==0.6.1